+++ CHANGES.md
@@ -8,6 +8,7 @@ Release Notes.
 * Support Tomcat thread pool metric collect.
 * Remove plugin for ServiceComb Java Chassis 0.x
 * Add Guava EventBus plugin.
+* Fix Dubbo 3.x plugin's tracing problem.
 * Support Druid Connection pool metrics collecting.
 * Support HikariCP Connection pool metrics collecting.
 * Support Dbcp2 Connection pool metrics collecting.